webapp/requirements.txt
django
django-rest-framework
numpy
pandas
gunicorn
sklearn



